pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "propertyhunter"
version = "0.1.0"
description = "Local-first property ingestion and search"
readme = "SPEC.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.1.0",
  "pydantic>=2.6.0",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  "streamlit>=1.31.0",
  "apscheduler>=3.10.0",
  "python-dotenv>=1.0.0",
]
